Package restructure and code review update
