+++ src/test/java/com/auth0/client/mgmt/UsersEntityTest.java
@@ -4,8 +4,8 @@ import com.auth0.client.mgmt.filter.FieldsFilter;
 import com.auth0.client.mgmt.filter.LogEventFilter;
 import com.auth0.client.mgmt.filter.PageFilter;
 import com.auth0.client.mgmt.filter.UserFilter;
-import com.auth0.json.mgmt.authenticationmethods.AuthenticationMethod;
-import com.auth0.json.mgmt.authenticationmethods.AuthenticationMethodsPage;
+import com.auth0.json.mgmt.users.authenticationmethods.AuthenticationMethod;
+import com.auth0.json.mgmt.users.authenticationmethods.AuthenticationMethodsPage;
 import com.auth0.json.mgmt.permissions.Permission;
 import com.auth0.json.mgmt.permissions.PermissionsPage;
 import com.auth0.json.mgmt.roles.RolesPage;
